Revert "all passed to ints"
This reverts commit db80ac1dc75b3c263ed89fc52b1c724d93a1ce96.

+++ GameClient/FoodBall.h
@@ -5,10 +5,10 @@ class FoodBall :
 	public Food
 {
 public:
-	FoodBall(int _id, sf::Vector2i pos);
+	FoodBall(int _id, sf::Vector2f pos);
 	~FoodBall();
 
 	sf::CircleShape circleShape;
-	int radius = 5;
+	float radius = 5.f;
 };
