+++ GameLib/Food.h
@@ -5,11 +5,11 @@ class Food
 {
 public:
 	Food();
-	Food(int _id, sf::Vector2i _pos);
+	Food(int _id, sf::Vector2f _pos);
 	~Food();
 
 	int id;
-	sf::Vector2i position;
+	sf::Vector2f position;
 	sf::Color color;
 
 private:
